backend/requirements.txt
fastapi
uvicorn
boto3
requests
python-multipart
python-dateutil
python-jose
